profile_agent/pyproject.toml
[project]
name = "profile-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fpdf>=1.7.2",
    "matplotlib>=3.10.3",
    "openai-agents>=0.0.17",
    "pymupdf>=1.26.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
]
